Generate function/class signatures using their params.
Currently, pyls returns the first line of docstring as function/class signature. For multi-line signatures, the returned value is incorrect. Also, when comments are interleaved with multi-line signatures, it will require some parsing to get the output correct.

This change generates the correct signatures by joining param descriptions with function names. This ensures that the returned signature is correct.