Add Lint Trigger
Pass a "trigger" string to the `pyls_lint` hook.  This allows linter
hook implementations to decide whether to run or not, based on what
triggered linting.  Useful for linters that are expensive to run,
so we don't want to run them on-change.

Since hook arguments are passed by keyword, this requires no changes
to existing local linter hook implementations.