setup.py: exclude all test packages
There is a subpackage in test. This subpackage was before this commit
installed although test package was ignored. Extending exclude to ignore
anything that starts with 'test' is enough to exclude any subpackage of
test.

This is causing an error when trying to package pyls in Gentoo. Gentoo
build system disutils eclass checks if there is some test package to be
installed and fails installation step if so. This fixes that.