Fix parent process being killed on Windows (#434)
Fixes issue #434

The `os.kill()` function being used to check for the existence of the
parent process causes the owning process to be killed on Windows
platforms.

This commit therefore introduces an alternative solution to check for
the parent process to be alive by using the Windows API

  types.windll.kernel32.OpenProcess()

An alternative would have been to introduce psutil library as another
dependency, which was judged not being worth.