Reduce redundant lint calls
Currently, whenever a file in the workspace changes all open documents are linted. This causes a lot of unnecessary lint calls.
With this PR changes to files not ending with `.py` or `.pyi` and changes that are already handled by the `didChange` event are ignored.

This issue is very problematic when using [`pyls-mypy'](https://github.com/tomv564/pyls-mypy/). `mypy` writes cache files after each lint which results in `didChangeWatchedFiles` to be triggered, resulting in a infinite lint loop.