Cache settings with lru_cache (#509)
`Config.settings()` takes a few milliseconds to run and reads config files from disk. This function runs multiple times during initialization and before every linting request. This means we can very effectively cache the results.

This PR uses `lru_cache` to cache the the result of `Config.settings()`. The cache is invalidated when the config is updated or a config file is changed.

Fixes #305