Changed pylint config search to search down from project root.
Normally, config files are searched for in ascending manner (i.e.
moving towards the root folder). However, in our configuration phase,
we don't know anything about any python files. Thus, root and document
path are identical, and this means that the config file can only be in
the project root directory.

The pylintrc file can be located in the current working
directory however, so we need to search *down* into subfolders.
Currently, we search at most 4 folders deep by default, but this can
be modified.