Allow usage of pylint via stdin
Mimics flake8 stdin usage.
This makes it possible for pylint to run on unsaved files
works on pylint>=2.5.0

I didn't notice a performance degradation in terms of running pylint itself.
Of course, there *is* a performance hit overall, as pylint will be run on every
change, and it can be fairly slow